requirements_backend.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# 기존 파이프라인 의존성
chromadb>=0.4.18
openai>=1.3.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
